Added minimal routing (intermediate update)
