+++ package.json
@@ -12,10 +12,12 @@
   "dependencies": {
     "@livekit/components-react": "^2.6.9",
     "@livekit/components-styles": "^1.1.4",
+    "antd": "^5.22.2",
     "axios": "^1.7.8",
     "livekit-client": "^2.7.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.0.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
